♻️ :: 공지사항 리스트 data class 중첩 클래스로 변경
