--- data/src/main/java/com/semicolon/data/remote/response/notice/NoticeListResponse.kt
@@ -1,7 +1,0 @@
-package com.semicolon.data.remote.response.notice
-
-import com.google.gson.annotations.SerializedName
-
-data class NoticeListResponse(
-    @SerializedName("notice_list") val noticeList: List<NoticeListValue>
-)
