--- data/src/main/java/com/semicolon/data/remote/response/notice/NoticeListValue.kt
@@ -1,10 +1,0 @@
-package com.semicolon.data.remote.response.notice
-
-import com.google.gson.annotations.SerializedName
-
-data class NoticeListValue (
-    @SerializedName("id") val noticeId: Int,
-    @SerializedName("title") val title: String,
-    @SerializedName("created_at") val createdAt: String,
-    @SerializedName("writer") val noticeWriter: NoticeWriter
-)
